refactor: service 내부 세부 패키지 제거 후 service 패키지 안에 통일
