+++ android/Bottari/data/src/main/java/com/bottari/data/service/FcmService.kt
@@ -1,4 +1,4 @@
-package com.bottari.data.service.fcm
+package com.bottari.data.service
 
 import com.bottari.data.model.remote.fcm.FcmTokenSaveRequest
 import retrofit2.Response
