+++ android/Bottari/data/src/main/java/com/bottari/data/source/remote/AlarmRemoteDataSourceImpl.kt
@@ -3,7 +3,7 @@ package com.bottari.data.source.remote
 import com.bottari.data.common.util.safeApiCall
 import com.bottari.data.model.remote.alarm.AlarmCreateRequest
 import com.bottari.data.model.remote.alarm.AlarmSaveRequest
-import com.bottari.data.service.alarm.AlarmService
+import com.bottari.data.service.AlarmService
 
 class AlarmRemoteDataSourceImpl(
     private val alarmService: AlarmService,
