+++ android/Bottari/data/src/main/java/com/bottari/data/source/remote/BottariItemRemoteDataSourceImpl.kt
@@ -3,7 +3,7 @@ package com.bottari.data.source.remote
 import com.bottari.data.common.util.safeApiCall
 import com.bottari.data.model.remote.bottari.item.ItemFetchResponse
 import com.bottari.data.model.remote.bottari.item.ItemsSaveRequest
-import com.bottari.data.service.bottari.item.BottariItemService
+import com.bottari.data.service.BottariItemService
 
 class BottariItemRemoteDataSourceImpl(
     private val bottariItemService: BottariItemService,
