+++ android/Bottari/data/src/main/java/com/bottari/data/source/remote/FcmRemoteDataSourceImpl.kt
@@ -2,7 +2,7 @@ package com.bottari.data.source.remote
 
 import com.bottari.data.common.util.safeApiCall
 import com.bottari.data.model.remote.fcm.FcmTokenSaveRequest
-import com.bottari.data.service.fcm.FcmService
+import com.bottari.data.service.FcmService
 
 class FcmRemoteDataSourceImpl(
     private val fcmService: FcmService,
