+++ android/Bottari/data/src/main/java/com/bottari/data/source/remote/ReportRemoteDataSourceImpl.kt
@@ -2,7 +2,7 @@ package com.bottari.data.source.remote
 
 import com.bottari.data.common.util.safeApiCall
 import com.bottari.data.model.remote.report.TemplateReportRequest
-import com.bottari.data.service.report.ReportService
+import com.bottari.data.service.ReportService
 
 class ReportRemoteDataSourceImpl(
     private val reportService: ReportService,
